+++ src/commands/measureunit.ts
@@ -1,6 +1,12 @@
 import Repzo from "repzo";
 import DataSet from "data-set-query";
-import { EVENT, Config, CommandEvent, Result } from "../types";
+import {
+  EVENT,
+  Config,
+  CommandEvent,
+  Result,
+  FailedDocsReport,
+} from "../types";
 import {
   _fetch,
   _create,
@@ -8,6 +14,7 @@ import {
   _delete,
   update_bench_time,
   updateAt_query,
+  set_error,
 } from "../util.js";
 
 interface QoyodUnit {
@@ -47,8 +54,8 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
       created: 0,
       updated: 0,
       failed: 0,
-      failed_msg: [],
     };
+    const failed_docs_report: FailedDocsReport = [];
 
     const qoyod_units: QoyodUnits = await get_qoyod_units(
       commandEvent.app.available_app.app_settings.serviceEndPoint,
@@ -128,12 +135,12 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
           const created_unit = await repzo.measureunit.create(body);
           result.created++;
         } catch (e: any) {
-          console.log("Create Measure Unit Failed >> ", e?.response, body);
-          result.failed_msg.push(
-            "Create Measure Unit Failed >> ",
-            e?.response,
-            body
-          );
+          // console.log("Create Measure Unit Failed >> ", e?.response, body);
+          failed_docs_report.push({
+            method: "create",
+            doc: body,
+            error_message: set_error(e),
+          });
           result.failed++;
         }
       } else {
@@ -151,12 +158,13 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
           );
           result.updated++;
         } catch (e: any) {
-          console.log("Update Measure Unit Failed >> ", e?.response, body);
-          result.failed_msg.push(
-            "Update Measure Unit Failed >> ",
-            e?.response,
-            body
-          );
+          // console.log("Update Measure Unit Failed >> ", e?.response, body);
+          failed_docs_report.push({
+            method: "update",
+            doc_id: repzo_unit?._id,
+            doc: body,
+            error_message: set_error(e),
+          });
           result.failed++;
         }
       }
@@ -170,13 +178,19 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
       bench_time_key,
       new_bench_time
     );
-    await commandLog.setStatus("success").setBody(result).commit();
+    await commandLog
+      .setStatus(
+        "success",
+        failed_docs_report.length ? failed_docs_report : null
+      )
+      .setBody(result)
+      .commit();
     return result;
   } catch (e: any) {
     //@ts-ignore
-    console.error(e?.response?.data);
+    console.error(e?.response?.data || e);
     await commandLog.setStatus("fail", e).commit();
-    throw e?.response;
+    throw e;
   }
 };