+++ changelog.md
@@ -20,6 +20,7 @@
 - divide create_payment to create_invoice_payment & create_receipt @maramalshen
 - [actions/create_refund] new action: refund @maramalshen
 - [actions/create_return_invoice] new action: credit notes @maramalshen
+- Update Error Logs in Actions & Commands @maramalshen
 
 ### Changed